requirements.txt
beautifulsoup4==4.8.2
mpmath==1.1.0
selenium==3.141.0
soupsieve==2.0
sympy==1.5.1
urllib3==1.25.8
python-dotenv==0.12.0
wheel
git+git://github.com/LarsVomMars/mathlib.git#egg=larsmathlib